refactor: adjust breed service into its own package
