+++ src/main/java/com/jpcchaves/adotar/application/service/breed/BreedServiceImpl.java
@@ -1,11 +1,13 @@
-package com.jpcchaves.adotar.application.service.impl.v1;
+package com.jpcchaves.adotar.application.service.breed;
 
 import com.jpcchaves.adotar.application.dto.pet.BreedDto;
-import com.jpcchaves.adotar.application.service.usecases.BreedService;
+import com.jpcchaves.adotar.application.service.breed.contracts.BreedService;
 import com.jpcchaves.adotar.application.utils.mapper.MapperUtils;
 import com.jpcchaves.adotar.domain.model.Breed;
 import com.jpcchaves.adotar.infra.repository.BreedRepository;
+
 import java.util.List;
+
 import org.springframework.stereotype.Service;
 
 @Service
@@ -14,7 +16,8 @@ public class BreedServiceImpl implements BreedService {
   private final BreedRepository breedRepository;
   private final MapperUtils mapper;
 
-  public BreedServiceImpl(BreedRepository breedRepository, MapperUtils mapper) {
+  public BreedServiceImpl(BreedRepository breedRepository,
+                          MapperUtils mapper) {
     this.breedRepository = breedRepository;
     this.mapper = mapper;
   }
