+++ src/main/java/com/jpcchaves/adotar/presentation/controller/v1/BreedController.java
@@ -1,7 +1,7 @@
 package com.jpcchaves.adotar.presentation.controller.v1;
 
 import com.jpcchaves.adotar.application.dto.pet.BreedDto;
-import com.jpcchaves.adotar.application.service.usecases.BreedService;
+import com.jpcchaves.adotar.application.service.breed.contracts.BreedService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -10,7 +10,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.websocket.server.PathParam;
+
 import java.util.List;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -35,27 +37,27 @@ public class BreedController {
       description =
           "Get a list of breeds by animal type (passing an animal type id)",
       responses = {
-        @ApiResponse(
-            description = "Success",
-            responseCode = "200",
-            content =
-                @Content(
-                    array =
-                        @ArraySchema(
-                            schema =
-                                @Schema(implementation = BreedDto.class)))),
-        @ApiResponse(
-            description = "Bad Request",
-            responseCode = "400",
-            content = @Content),
-        @ApiResponse(
-            description = "Unauthorized",
-            responseCode = "401",
-            content = @Content),
-        @ApiResponse(
-            description = "Internal Error",
-            responseCode = "500",
-            content = @Content),
+          @ApiResponse(
+              description = "Success",
+              responseCode = "200",
+              content =
+              @Content(
+                  array =
+                  @ArraySchema(
+                      schema =
+                      @Schema(implementation = BreedDto.class)))),
+          @ApiResponse(
+              description = "Bad Request",
+              responseCode = "400",
+              content = @Content),
+          @ApiResponse(
+              description = "Unauthorized",
+              responseCode = "401",
+              content = @Content),
+          @ApiResponse(
+              description = "Internal Error",
+              responseCode = "500",
+              content = @Content),
       })
   public ResponseEntity<List<BreedDto>> findAllByAnimalType(
       @PathParam("animalTypeId") Long animalTypeId) {
